fixed and refactored vr tests
